\name{fpScale}
\alias{fpScale}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Internal function for shifting and rescaling a fractional polynomial covariate}
\description{
  A copy of mfp:::fp.scale, see its description.
}
\usage{
fpScale(x, scaling = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{covariate}
  \item{scaling}{should scaling take place?}
}

\keyword{math}
\keyword{internal}