\name{fpScale}
\alias{fpScale}

\title{Internal function for shifting and rescaling a fractional polynomial covariate}
\description{
  This function is a copy of mfp:::fp.scale, see its description.
}
\usage{
fpScale(x, scaling = TRUE)
}

\arguments{
  \item{x}{covariate}
  \item{scaling}{should scaling take place?}
}

\keyword{math}
\keyword{internal}